homework 3. add typescript
