+++ package.json
@@ -4,10 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "babel hw-1 -d lib",
-    "task1": "nodemon --exec babel-node hw-1/task1.js ",
-    "task2": "nodemon --exec babel-node hw-1/task2.js ",
-    "task2-1": "nodemon --exec babel-node hw-2/server.js "
+    "build": "npx tsc",
+    "start": "node dist/server.js",
+    "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/hw-2/server.js\""
   },
   "repository": {
     "type": "module",
@@ -24,10 +23,16 @@
     "@babel/core": "^7.20.5",
     "@babel/node": "^7.20.5",
     "@babel/preset-env": "^7.20.2",
-    "eslint": "^8.29.0",
+    "@types/express": "^4.17.15",
+    "@types/node": "^18.11.18",
+    "@typescript-eslint/eslint-plugin": "^5.48.0",
+    "@typescript-eslint/parser": "^5.48.0",
+    "concurrently": "^7.6.0",
+    "eslint": "^8.31.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.26.0",
-    "nodemon": "^2.0.20"
+    "nodemon": "^2.0.20",
+    "typescript": "^4.9.4"
   },
   "dependencies": {
     "csvtojson": "^2.0.10",
