+++ src/loaders/index.ts
@@ -1,8 +1,8 @@
 import postgressLoader from './postgress';
 
 const loaders = () => {
-  const pgConnection = postgressLoader();
-  console.log('DB Initialized', pgConnection);
+    const pgConnection = postgressLoader();
+    console.log('DB Initialized', pgConnection);
 };
 
 export default loaders;
